Move files, reconstitute files. In order to make preparations for pr
