+++ iOS/AudioPalyerWights/AppIntent.swift
@@ -1,0 +1,18 @@
+//
+//  AppIntent.swift
+//  AudioPalyerWights
+//
+//  Created by 四月初一茶染绮良 on 2025/3/7.
+//
+
+import WidgetKit
+import AppIntents
+
+struct ConfigurationAppIntent: WidgetConfigurationIntent {
+    static var title: LocalizedStringResource { "Configuration" }
+    static var description: IntentDescription { "This is an example widget." }
+
+    // An example configurable parameter.
+    @Parameter(title: "Favorite Emoji", default: "😃")
+    var favoriteEmoji: String
+}
