+++ iOS/AudioPalyerWights/AudioPalyerWightsBundle.swift
@@ -1,0 +1,16 @@
+//
+//  AudioPalyerWightsBundle.swift
+//  AudioPalyerWights
+//
+//  Created by 四月初一茶染绮良 on 2025/3/7.
+//
+
+import WidgetKit
+import SwiftUI
+
+@main
+struct AudioPalyerWightsBundle: WidgetBundle {
+    var body: some Widget {
+        AudioPalyerWights()
+    }
+}
