+++ iOS/AudioPlayerColorful/Assets.xcassets/AppIcon.appiconset/Contents.json
