+++ iOS/AudioPlayerColorful/Assets.xcassets/Contents.json
