+++ iOS/AudioPlayerColorful/Setting/Settings.bundle/Root.plist
