+++ iOS/AudioPlayerColorful/Setting/Settings.bundle/en.lproj/Root.strings
