+++ iOS/AudioPlayerColorful/Views/MoreView.swift
@@ -1,0 +1,21 @@
+//
+//  MoreView.swift
+//  AudioPlayerColorful
+//
+//  Created by 四月初一茶染绮良 on 2025/3/8.
+//
+
+import SwiftUI
+
+// MARK: More View
+struct MoreView: View
+{
+    var body: some View
+    {
+        NavigationStack
+        {
+            Text("更多内容")
+                .navigationTitle("更多")
+        }
+    }
+}
