+++ iOS/AudioPlayerColorful/Views/ViewSettings/HelpView.swift
@@ -1,0 +1,15 @@
+//
+//  HelpView.swift
+//  AudioPlayerColorful
+//
+//  Created by 四月初一茶染绮良 on 2025/3/8.
+//
+
+import SwiftUI
+
+// MARK: HelpView
+struct HelpView: View {
+    var body: some View {
+        Text("帮助中心")
+    }
+}
