+++ iOS/AudioPlayerColorful/Views/ViewSettings/PrivacyView.swift
@@ -1,0 +1,16 @@
+//
+//  PrivacyView.swift
+//  AudioPlayerColorful
+//
+//  Created by 四月初一茶染绮良 on 2025/3/8.
+//
+
+import SwiftUI
+
+struct PrivacyView: View
+{
+    var body: some View
+    {
+        Text("隐私设置")
+    }
+}
